SCRUM-2 added throw IllegalArgumentException to methods in UserService

+++ src/main/java/pl/galushop/GaluShop/service/UserService.java
@@ -16,11 +16,17 @@ public class UserService {
     private final PasswordEncoder passwordEncoder;
 
     public User getUser(Long userId){
+        if (userId == null || userId < 1) {
+            throw new IllegalArgumentException("Invalid user ID: " + userId);
+        }
         return userRepository.findById(userId)
                 .orElseThrow(() -> new UsernameNotFoundException("User with ID " + userId + " not found"));
     }
 
     public void deleteUser(Long userId){
+        if (userId == null || userId < 1) {
+            throw new IllegalArgumentException("Invalid user ID: " + userId);
+        }
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new UsernameNotFoundException("User with ID " + userId + " not found"));
         userRepository.delete(user);
@@ -28,6 +34,9 @@ public class UserService {
 
     @Transactional
     public void updateUser(UserRequest userRequest){
+        if(userRequest.getUserId() == null || userRequest.getUserId() < 0){
+            throw new IllegalArgumentException("Invalid user request: missing or incorrect user ID");
+        }
         User existingUser = userRepository.findById(userRequest.getUserId())
                 .orElseThrow(() -> new UsernameNotFoundException("User with ID " + userRequest.getUserId() + " not found"));
